requirements.txt
Django==4.2.13
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0